refactor folder structure, add strategy subdir
