+++ src/modrover/__init__.py
@@ -1,2 +1,1 @@
 from .main import Rover
-from .modelhub import ModelHub
