+++ src/modrover/strategies/__init__.py
@@ -1,0 +1,7 @@
+from modrover.strategies.full_explore import FullExplore
+
+strategy_type_dict = {
+    "full": FullExplore,
+    "down": DownExplore,
+    "up": UpExplore,
+}
